Add code coverage stats
Electron-mocha makes things like this a little difficult.  Luckily,
@inukshuk in
https://github.com/jprichardson/electron-mocha/issues/19#issuecomment-227868475
figures out a way to get around it. It's a bit of effort but does work
in the end.

With any luck an npm module will come out with these clues embedded in
it and the coverage.js can be removed.