Initial travis/coveralls integration
Don't build vN.N.N (release) branches. Only notify after all bits of a
matrix build.

The notification goes to the IRC channel as a notice from
off-channel. +n needs to be set to allow that.

The ignores are allowing for 100% coverage to be reported. I've only
used them where automated testing would be very difficult.

This also adds the glob and sinon modules that are necessary for tests
to work properly.